Archetype creation is working now. An archetype is just a unique structure per collection of component types

+++ Source/EngineCore/DirectoryLocations.cpp
@@ -7,7 +7,7 @@ String EngineExeFullPath()
 	HMODULE hCurrentModule;
 	GetModuleHandleEx(GET_MODULE_HANDLE_EX_FLAG_FROM_ADDRESS | GET_MODULE_HANDLE_EX_FLAG_UNCHANGED_REFCOUNT, (LPCSTR)&EngineAssetPath, &hCurrentModule);
 	tchar pathResult[512] = "";
-	GetModuleFileName(hCurrentModule, pathResult, ArraySize(pathResult));
+	GetModuleFileName(hCurrentModule, pathResult, (DWORD)ArraySize(pathResult));
 	size_t pathLen = Strlen(pathResult);
 	if (pathLen > 0)
 	{
